+++ src/components/Table/index.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import {
-  useTable, useSortBy,
+  useTable, useSortBy, usePagination,
   type UseTableOptions, Row, ColumnInstance, UseSortByColumnProps
 } from 'react-table';
 
@@ -12,6 +12,7 @@ import { NonIdealState } from '../NonIdealState';
 import TableRow from './TableRow';
 import { IconSortable } from './IconSortable';
 import image from '../Icon/icons/IconBoxNoData/empty-box-no-data.svg';
+import { Pagination } from '../Pagination';
 
 
 const styles = {
@@ -37,7 +38,13 @@ const styles = {
   noDataText: css`
     margin-top: 16px;
     color: rgba(0, 0, 0, 0.25);
- `
+ `,
+  pagination: css`
+    margin-top: 40px;
+    display: flex;
+    justify-content: flex-end;
+    margin-right: 16px;
+  `
 };
 
 
@@ -46,6 +53,7 @@ export type RowProps = {
   codeClassName?: string,
   codeRowKey?: string,
   onClickCodeRow?: (row: Row) => void;
+  pagination?: boolean;
 }
 
 type TableProps = UseTableOptions & RowProps;
@@ -64,7 +72,9 @@ function getSortDirection(isSortedDesc?: boolean) {
 }
 
 export function Table(props: TableProps) {
-  const { rowClassName, codeClassName, columns = [], data = [] } = props;
+  const {
+    rowClassName, codeClassName, columns = [], data = [], pagination
+  } = props;
 
   const {
     getTableProps,
@@ -72,15 +82,22 @@ export function Table(props: TableProps) {
     headerGroups,
     prepareRow,
     rows,
-    visibleColumns
+    page,
+    visibleColumns,
+
+    gotoPage,
+    setPageSize,
+    state: { pageIndex, pageSize }
   } = useTable(
     {
       columns,
       data
     },
-    useSortBy
+    useSortBy,
+    usePagination
   );
 
+  const dataRows = pagination ? page : rows;
   return (
     <>
       <table {...getTableProps()} className={cx(styles.table)}>
@@ -103,7 +120,7 @@ export function Table(props: TableProps) {
           ))}
         </thead>
         <tbody {...getTableBodyProps()}>
-          {rows.map(row => {
+          {dataRows.map(row => {
             prepareRow(row);
             return (
               <TableRow
@@ -127,6 +144,16 @@ export function Table(props: TableProps) {
           )}
         </tbody>
       </table>
+      {pagination && rows.length > 0 && <div className={styles.pagination}>
+        <Pagination
+          page={pageIndex}
+          pageSize={pageSize}
+          items={rows.length}
+          onPageChange={gotoPage}
+          setPageSize={setPageSize}
+          showTotal
+        />
+      </div>}
     </>
   );
 }