+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - IconPlay
   - IconStop
   - IconTask
+- Added new component Pagination
+- Usage Pagination in table  
 
 - Improve Icon's readme.